Improve management of outputs encoding.
The encoding of outputs is defined by the '_decode' attribute (its default value
is 'utf-8'). If the attribute is defined, outputs (stdout and stderr) will be of
type 'unicode' for python2 and 'str' for python3 with the given value of the
attribute used for decoding. If '_decode' value is None, python2 str/python3
bytes are returned.

'_decode' value can be set with the 'set_control' method of the 'set_controls'
context manager.