Add a method for itering over stdout and stderr (by returning a generator).
The 'iter' method has been defined for both 'Local' and 'Remote' and implements
a generator for itering over stdout and stderr lines. Each iteration is a tuple
of two elements with the firsh element indicating whether the line comes from
stdout and stderr and the second element the line value. The last tuple contains
the status of the command.